Added CSS, IMG
CSS: Added a colour theme using HEX code in the root, default margins and padding added, Playing around with styling thinking it would be better if we had a reference website to go from so we have something to work with.
HTML: Added a Google font that is popular and easy to read, classes added to h1 and h2 elements thus not targeting all but the ones I wanted.
IMG: Added 2 images one for the background(can change) and another one not for use but for us to kind of see what the colour scheme would look like.