Search bar border
CSS: Search bar border had a default style from the user agent sheet but removed it with (border: none;) and now its not expanding.
Figured out why the search bar was moving and used the property of flex: 1; to fix this and what it does is give each child a 1 for 1 for 1 spacing meaning it will not exceed 33% of the div container.